# perf(zip)!: don't access the filesystem from wasm