Depends on https://github.com/yarnpkg/berry/pull/5179.

**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

This PR makes `ZipFS` read the archive directly inside the constructor from Node and passing a buffer to libzip instead of making libzip read it with a shimmed FS.

Consequently, `ZipFS` now only accesses the filesystem twice for each archive (first time in the constructor and second time in `saveAndClose`) and therefore doesn't have to keep an fd handle to the archive while it is open.
This means that now:
- all "out of file descriptors" errors should finally be gone
- the only limit to how many archives can be open at the same time is memory.


This could have the consequence of allowing us to increase the fetcher and linker concurrency limits, but I haven't tested it yet.

Fixes https://github.com/yarnpkg/berry/issues/5172.
Fixes https://github.com/yarnpkg/berry/issues/4488.

**How did you fix it?**
<!-- A detailed description of your implementation. -->

This has the advantage of greatly reducing the bundle size since libzip doesn't have to bundle the FS shims anymore:

|                       |  Before |  After  |  Change |
|-----------------------|:-------:|:-------:|:-------:|
| Bundle size (bytes)   | 2800175 | 2693300 |  -3.82% |
| PnP Hook size (bytes) |  162041 |  136565 | -15.72% |

It also seems to bring performance improvements on `yarn install` (though earlier I saw some mixed results when benchmarking the time of accessing many archives so if anybody wants to run more benchmarks feel free, but I'd say that it's good enough since I could consistently reproduce the Yarn speedup):

```
Reflink aren't supported! Installs may be quite slower than necessary
Testing install-full-cold
Benchmark 1: node ./before.cjs install --mode skip-build
  Time (mean ± σ):     20.536 s ±  0.848 s    [User: 41.778 s, System: 2.380 s]
  Range (min … max):   19.821 s … 22.663 s    10 runs
 
Benchmark 2: node ./after.cjs install --mode skip-build
  Time (mean ± σ):     20.170 s ±  0.705 s    [User: 41.661 s, System: 2.380 s]
  Range (min … max):   19.583 s … 22.006 s    10 runs
 
Summary
  'node ./after.cjs install --mode skip-build' ran
    1.02 ± 0.06 times faster than 'node ./before.cjs install --mode skip-build'
Testing install-cache-only
Benchmark 1: node ./before.cjs install --mode skip-build
  Time (mean ± σ):     10.256 s ±  1.375 s    [User: 12.546 s, System: 1.329 s]
  Range (min … max):    9.248 s … 12.956 s    10 runs
 
Benchmark 2: node ./after.cjs install --mode skip-build
  Time (mean ± σ):      9.310 s ±  0.175 s    [User: 12.402 s, System: 1.316 s]
  Range (min … max):    9.078 s …  9.595 s    10 runs
 
Summary
  'node ./after.cjs install --mode skip-build' ran
    1.10 ± 0.15 times faster than 'node ./before.cjs install --mode skip-build'
Testing install-cache-and-lock
Benchmark 1: node ./before.cjs install --mode skip-build
  Time (mean ± σ):      2.239 s ±  0.015 s    [User: 3.452 s, System: 0.514 s]
  Range (min … max):    2.206 s …  2.257 s    10 runs
 
Benchmark 2: node ./after.cjs install --mode skip-build
  Time (mean ± σ):      2.112 s ±  0.043 s    [User: 3.325 s, System: 0.485 s]
  Range (min … max):    2.027 s …  2.169 s    10 runs
 
Summary
  'node ./after.cjs install --mode skip-build' ran
    1.06 ± 0.02 times faster than 'node ./before.cjs install --mode skip-build'
Testing install-ready
Benchmark 1: node ./before.cjs add dummy-pkg@link:./dummy-pkg --mode skip-build
  Time (mean ± σ):     942.9 ms ±   7.2 ms    [User: 1253.3 ms, System: 100.4 ms]
  Range (min … max):   931.6 ms … 953.5 ms    10 runs
 
Benchmark 2: node ./after.cjs add dummy-pkg@link:./dummy-pkg --mode skip-build
  Time (mean ± σ):     823.4 ms ±  16.4 ms    [User: 1145.1 ms, System: 83.3 ms]
  Range (min … max):   802.9 ms … 861.7 ms    10 runs
 
Summary
  'node ./after.cjs add dummy-pkg@link:./dummy-pkg --mode skip-build' ran
    1.15 ± 0.02 times faster than 'node ./before.cjs add dummy-pkg@link:./dummy-pkg --mode skip-build'
```

**Checklist**
<!--- Don't worry if you miss something, chores are automatically tested. -->
<!--- This checklist exists to help you remember doing the chores when you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [X] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).

<!-- See https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check -i` -->
- [X] I have set the packages that need to be released for my changes to be effective.

<!-- The "Testing chores" workflow validates that your PR follows our guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR might be missing. -->
- [X] I will check that all automated PR checks pass before the PR gets reviewed.