@arcanis 2023-01-06
That seems reasonable. It's also what #5145 does.