@belgattitude 2023-01-21
@paul-soporan I sometimes bench pnpm and yarn... Lately I saw a perf boost from yarn 4.0.0-rc.35 and 36 (more than 30% speedup) with a monorepo (nextjs, vite, esbuild, vitest...)

Not sure if it's linked but wanted to let you know.

The benchs:

- https://github.com/belgattitude/compare-package-managers#ci-with-cache
- https://github.com/belgattitude/compare-package-managers/blob/main/.github/workflows/ci-install-benchmark.yml

![image](https://user-images.githubusercontent.com/259798/213887285-3d18b45b-05c3-49be-95cb-d1a1120d3ceb.png)

Note that it's the first time the install was faster than pnpm (YARN_COMPRESSION_LEVEL=0 and dedupe)... pnpm does download more (peer-deps handling differs so difficult to compare)

Speedup appreciated ♻️🌳