@MarcCelani-at 2023-02-09
@merceyz I only read the summary here, didn't dig into the details, but can you help me understand why we can't backport this? We really need this in our monorepo :) IDK if there is a timeline for yarn 4 or not...